refactor: P3-AUDIT - 抽离reader句子装饰调度
