+++ src/ui/conversations/views/ArticleReaderView.tsx
@@ -6,19 +6,27 @@ import { readerPrefsToCssVars } from '@services/protocols/reader-prefs';
 import { buildSentences, type ReaderTtsSentence } from '@services/reader/tts/reader-tts-engine';
 import type { DetailViewSharedProps } from '@ui/conversations/views/detail-view-props';
 import {
-  collectReaderSentenceTextSegments,
   findReaderSentenceIndexFromTarget,
-  findReaderSentenceRangeByIndex,
-  pickFirstVisibleSentenceIndex,
-  type ReaderSentenceCandidate,
-  type ReaderSentenceRectLike,
 } from '@ui/reader/reader-sentence-dom';
-import {
-  publishReaderPerformanceStats,
-  readReaderPerformanceClock,
-} from '@ui/reader/reader-performance-debug';
 import { useArticleOutlineMinimap } from '@ui/reader/ArticleOutlineMinimap';
 import type { ReaderOutlineDomEntry } from '@ui/reader/article-outline-dom';
+import {
+  clearReaderSentenceDecorations,
+  decorateReaderSentenceSpans,
+  decorateReaderSentenceSpansProgressively,
+  readFirstVisibleReaderSentenceIndex,
+  READER_CURRENT_SENTENCE_CLASS,
+  READER_EAGER_DECORATION_SENTENCE_LIMIT,
+  READER_EAGER_DECORATION_SOURCE_LENGTH_LIMIT,
+  READER_INITIAL_SENTENCE_DECORATION_BATCH_SIZE,
+  READER_REDECORATE_SETTLE_MS,
+  READER_SENTENCE_DECORATION_PENDING,
+  READER_SENTENCE_DECORATION_READY,
+  READER_SENTENCE_DECORATION_STATUS_ATTR,
+  READER_SENTENCE_INDEX_ATTR,
+  READER_SENTENCE_SOURCE_ATTR,
+} from '@ui/reader/reader-sentence-decoration';
+import { publishReaderPerformanceStats, readReaderPerformanceClock } from '@ui/reader/reader-performance-debug';
 import { ReaderToolbar } from '@ui/reader/ReaderToolbar';
 import { useReaderNarration } from '@viewmodels/reader/useReaderNarration';
 import { useReaderPrefs } from '@viewmodels/reader/useReaderPrefs';
@@ -51,159 +59,6 @@ const READER_COLUMN_STYLE: CSSProperties = { maxWidth: 'var(--reader-content-wid
 const READER_SHELL_CLASS = 'tw-flex tw-w-full tw-items-start tw-gap-4';
 const READER_MAIN_CLASS = 'tw-min-w-0 tw-flex-1 tw-max-w-full';
 const READER_RAIL_CLASS = 'tw-flex-none tw-shrink-0 tw-self-start';
-
-const READER_SENTENCE_SOURCE_ATTR = 'data-reader-sentence-source';
-const READER_SENTENCE_INDEX_ATTR = 'data-reader-sentence-index';
-const READER_SENTENCE_DECORATION_STATUS_ATTR = 'data-reader-sentence-decoration';
-const READER_SENTENCE_DECORATION_PENDING = 'pending';
-const READER_SENTENCE_DECORATION_READY = 'ready';
-const READER_SENTENCE_CLASS = 'reader-tts-sentence';
-const READER_CURRENT_SENTENCE_CLASS = 'reader-tts-sentence-active';
-const READER_INITIAL_SENTENCE_DECORATION_BATCH_SIZE = 48;
-const READER_REDECORATE_SETTLE_MS = 180;
-const READER_EAGER_DECORATION_SENTENCE_LIMIT = 400;
-const READER_EAGER_DECORATION_SOURCE_LENGTH_LIMIT = 16000;
-
-function unwrapReaderSentenceSpan(span: HTMLElement): void {
-  const parent = span.parentNode;
-  if (!parent) return;
-  while (span.firstChild && span.parentNode === parent) {
-    parent.insertBefore(span.firstChild, span);
-  }
-  if (span.parentNode === parent) {
-    parent.removeChild(span);
-  }
-}
-
-function clearReaderSentenceDecorations(root: HTMLElement): void {
-  const spans = Array.from(root.querySelectorAll<HTMLElement>(`[${READER_SENTENCE_INDEX_ATTR}]`));
-  spans.forEach((span) => unwrapReaderSentenceSpan(span));
-  root.removeAttribute(READER_SENTENCE_SOURCE_ATTR);
-  root.removeAttribute(READER_SENTENCE_DECORATION_STATUS_ATTR);
-  root.normalize();
-}
-
-function decorateReaderSentenceRangeBatch(
-  root: HTMLElement,
-  sentences: ReaderTtsSentence[],
-  startIndex: number,
-  endIndex: number,
-): number {
-  const startedAt = readReaderPerformanceClock();
-  const segments = collectReaderSentenceTextSegments(root);
-  for (let index = endIndex - 1; index >= startIndex; index -= 1) {
-    const sentence = sentences[index];
-    const range = findReaderSentenceRangeByIndex(root, sentences, index, segments);
-    if (!range) continue;
-
-    const fragment = range.extractContents();
-    if (!fragment.childNodes.length) continue;
-
-    const span = (root.ownerDocument ?? globalThis.document)?.createElement('span');
-    if (!span) continue;
-    span.className = READER_SENTENCE_CLASS;
-    span.setAttribute(READER_SENTENCE_INDEX_ATTR, String(sentence.index));
-    span.appendChild(fragment);
-    range.insertNode(span);
-  }
-  return Math.max(0, readReaderPerformanceClock() - startedAt);
-}
-
-function decorateReaderSentenceSpans(root: HTMLElement, source: string, sentences: ReaderTtsSentence[]): void {
-  const doc = root.ownerDocument ?? globalThis.document;
-  if (!doc) return;
-  if (!sentences.length) {
-    root.setAttribute(READER_SENTENCE_SOURCE_ATTR, source);
-    root.setAttribute(READER_SENTENCE_DECORATION_STATUS_ATTR, READER_SENTENCE_DECORATION_READY);
-    return;
-  }
-  const batchDuration = decorateReaderSentenceRangeBatch(root, sentences, 0, sentences.length);
-  root.setAttribute(READER_SENTENCE_SOURCE_ATTR, source);
-  root.setAttribute(READER_SENTENCE_DECORATION_STATUS_ATTR, READER_SENTENCE_DECORATION_READY);
-  publishReaderPerformanceStats((current) => ({
-    ...current,
-    decorateMode: 'sync',
-    decoratePasses: current.decoratePasses + 1,
-    decorateBatches: current.decorateBatches + 1,
-    decorateLastBatchDurationMs: batchDuration,
-    decorateLastDurationMs: batchDuration,
-    decorateTotalDurationMs: current.decorateTotalDurationMs + batchDuration,
-    decorateMaxBatchDurationMs: Math.max(current.decorateMaxBatchDurationMs, batchDuration),
-  }));
-}
-
-function decorateReaderSentenceSpansProgressively(
-  root: HTMLElement,
-  source: string,
-  sentences: ReaderTtsSentence[],
-  requestFrame: (cb: FrameRequestCallback) => number,
-  cancelFrame: ((id: number) => void) | undefined,
-  onProgress: () => void,
-): () => void {
-  let cancelled = false;
-  let rafId = 0;
-  let nextIndex = 0;
-
-  root.setAttribute(READER_SENTENCE_SOURCE_ATTR, source);
-  root.setAttribute(READER_SENTENCE_DECORATION_STATUS_ATTR, READER_SENTENCE_DECORATION_PENDING);
-
-  const step = () => {
-    if (cancelled) return;
-    const batchEnd = Math.min(nextIndex + READER_INITIAL_SENTENCE_DECORATION_BATCH_SIZE, sentences.length);
-    const batchDuration = decorateReaderSentenceRangeBatch(root, sentences, nextIndex, batchEnd);
-    nextIndex = batchEnd;
-    publishReaderPerformanceStats((current) => ({
-      ...current,
-      decorateMode: 'progressive',
-      decorateBatches: current.decorateBatches + 1,
-      decorateLastBatchDurationMs: batchDuration,
-      decorateTotalDurationMs: current.decorateTotalDurationMs + batchDuration,
-      decorateMaxBatchDurationMs: Math.max(current.decorateMaxBatchDurationMs, batchDuration),
-    }));
-    if (nextIndex >= sentences.length) {
-      root.setAttribute(READER_SENTENCE_DECORATION_STATUS_ATTR, READER_SENTENCE_DECORATION_READY);
-      publishReaderPerformanceStats((current) => ({
-        ...current,
-        decoratePasses: current.decoratePasses + 1,
-        decorateLastDurationMs: current.decorateLastDurationMs + 0,
-      }));
-      onProgress();
-      return;
-    }
-    onProgress();
-    rafId = requestFrame(() => step());
-  };
-
-  step();
-  return () => {
-    cancelled = true;
-    if (rafId !== 0) cancelFrame?.(rafId);
-  };
-}
-
-function readReaderSentenceViewportRect(root: HTMLElement): ReaderSentenceRectLike {
-  const view = root.ownerDocument?.defaultView ?? globalThis.window ?? null;
-  const height = Number(view?.innerHeight);
-  if (!Number.isFinite(height) || height <= 0) return { top: 0, bottom: 0 };
-  return { top: 0, bottom: height };
-}
-
-function readReaderSentenceCandidates(root: HTMLElement): ReaderSentenceCandidate[] {
-  return Array.from(root.querySelectorAll<HTMLElement>(`[${READER_SENTENCE_INDEX_ATTR}]`))
-    .map((element) => {
-      const rect = element.getBoundingClientRect();
-      const index = Number(element.getAttribute(READER_SENTENCE_INDEX_ATTR));
-      return {
-        index: Number.isFinite(index) ? Math.trunc(index) : -1,
-        rect: { top: rect.top, bottom: rect.bottom },
-      } satisfies ReaderSentenceCandidate;
-    })
-    .filter((candidate) => candidate.index >= 0);
-}
-
-function readFirstVisibleSentenceIndex(root: HTMLElement): number {
-  return pickFirstVisibleSentenceIndex(readReaderSentenceCandidates(root), readReaderSentenceViewportRect(root));
-}
 
 /**
  * ArticleReaderView renders article / video conversations.
@@ -554,7 +409,7 @@ export function ArticleReaderView({
     if (!features.narration) return 0;
     const root = narrationRootRef.current;
     if (!root) return 0;
-    return readFirstVisibleSentenceIndex(root);
+    return readFirstVisibleReaderSentenceIndex(root);
   }, [features.narration]);
 
   const toolbarNarration = useMemo(
