+++ src/ui/i18n/locales/zh.ts
@@ -126,9 +126,6 @@ export const zh: { [K in TranslationKey]: string } = {
   readerNarrationPause: '暂停',
   readerNarrationStop: '停止',
   readerTextPreset: '预设',
-  readerPresetMedium: 'Medium',
-  readerPresetNotion: 'Notion',
-  readerPresetBook: 'Book',
   readerTextFont: '字体',
   readerFontAria: '阅读字体',
   readerFontSerif: '衬线',
